A little refactoring
